# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/users/{id}/suspend": {
      "put": {
        "operationId": "AdminUsersController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminUsers"
        ]
      }
    },
...
    "/api/v1/admin/customers/{id}": {
      "get": {
        "operationId": "AdminCustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCustomers"
        ]
      }
    },
    "/api/v1/admin/customers/{id}/suspend": {
      "put": {
        "operationId": "AdminCustomersController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCustomers"
        ]
      }
    },
    "/api/v1/admin/customers/{id}/unsuspend": {
      "put": {
        "operationId": "AdminCustomersController_unsuspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCustomers"
        ]
      }
    },
    "/api/v1/admin/customers/{id}/reset-password": {
      "post": {
        "operationId": "AdminCustomersController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCustomers"
        ]
      }
    },
    "/api/v1/admin/customers/{id}/notes": {
      "put": {
        "operationId": "AdminCustomersController_updateNotes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCustomers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Aurora API",
    "description": "Aurora E-Commerce API Endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuspendUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}